requirements.txt
discord.py==1.7.3
twython==3.9.1
discord-py-slash-command==2.3.2